+++ utility/ObjectPool.hpp
@@ -2,7 +2,7 @@
  * @Author: AClolinta AClolinta@gmail.com
  * @Date: 2023-04-03 12:06:35
  * @LastEditors: AClolinta AClolinta@gmail.com
- * @LastEditTime: 2023-06-16 11:45:49
+ * @LastEditTime: 2023-06-19 09:59:10
  * @FilePath: /TinyWebServer/utility/ObjectPool.hpp
  * @Description:  对象池
  *  */
@@ -51,7 +51,7 @@ ObjectPool<T>::~ObjectPool() {
 template <class T>
 void ObjectPool<T>::Init(size_t max) {
     AutoLock lock(&m_mutex);
-    for (int i = 0; i < max; ++i) {
+    for (size_t i = 0; i < max; ++i) {
         T* ptr = new T();
         m_pool.push_back(ptr);
     }
